pyproject.toml
[tool.poetry]
name = "horsona"
version = "0.1.0"
description = ""
readme = "README.md"
authors = [
  "Synthbot Anon <synthbot.anon@gmail.com>"
]

[tool.poetry.dependencies]
python = "^3.11"
cerebras_cloud_sdk = "^1.0.0"
fireworks-ai = "^0.15.0"
groq = "^0.9.0"
python-dotenv = "^1.0.1"
pyyaml = "^6.0.2"
together = "^1.2.7"
transformers = "^4.44.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.24.0"



[tool.poetry.group.compute-cpu.dependencies]
torch = {version = "^2.4.0+cpu", source = "pytorch-cpu"}

[[tool.poetry.source]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

